Refactor into single package
