--- ui/constants.py
@@ -1,1 +1,0 @@
-OPTION_WRITE_RAW_TEMPLATE = "pp_write_raw_template"
